:recycle: clean up getDefinitionAtPosition decorator
