+++ ts-plugin/src/decorators/getScriptKind.ts
@@ -1,11 +1,11 @@
-import { createDecorator } from '../factories'
+import { createHostDecorator } from '../factories'
 import { isSolidity } from '../utils'
 
 /**
  * Decorate `LangaugeServerHost.getScriptKind` to return TS type for `.sol` files
  * This lets the ts-server expect `.sol` files to resolve to `.d.ts` files in `resolveModuleNameLiterals`
  */
-export const getScriptKindDecorator = createDecorator((createInfo, ts) => {
+export const getScriptKindDecorator = createHostDecorator((createInfo, ts) => {
 	return {
 		getScriptKind: (fileName) => {
 			if (isSolidity(fileName)) {
